PLT-8188: Add scopes to permissions.
